Demo: Don't return an error code (to the OS)
... on failure to open file.
As the demo is demonstrating the existence of the problem, there is no
real meaning to returning an error code. The printed message is enough
to make the point.

Also, this helps AppVeyor pass successfully (and it should, as this is
not a build or test failure), without having to resort to convoluted
hacks.